+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Add expo-router route for `auth/callback` deep link (fixes "unmatched route" error)
+
 ## [1.5.1] - 2026-01-05
 
 ### Fixed